mcp_server/requirements.txt
# MCP Server Dependencies
mcp>=1.0.0
pymongo>=4.6.1
mongoengine>=0.27.0
python-dotenv>=1.0.0
